refactor: split onboard guided config
